manage userId in expense crud

+++ Controllers/ExpenseController.cs
@@ -5,6 +5,7 @@ using Microsoft.EntityFrameworkCore;
 using ExpenseTrack.Data;
 using ExpenseTrack.Models;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using System.Security.Claims;
 
 namespace ExpenseTrack.Controllers
 {
@@ -20,14 +21,24 @@ namespace ExpenseTrack.Controllers
         // GET: Expense
         public IActionResult Index(DateTime? filterDate)
         {
-            var expenses = _context.Expenses.Include(e => e.Category).ToList();
-			if (filterDate.HasValue)
-			{
-				expenses = expenses.Where(e => e.Date.Date == filterDate.Value.Date).ToList();
-			}
-			return View(expenses);
+            // Get the ID of the currently logged-in user
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            // Filter expenses based on the user ID
+            var expenses = _context.Expenses
+                .Include(e => e.Category)
+                .Where(e => e.UserId == userId)
+                .ToList();
+
+            if (filterDate.HasValue)
+            {
+                expenses = expenses.Where(e => e.Date.Date == filterDate.Value.Date).ToList();
+            }
+
+            return View(expenses);
         }
-		public IActionResult AddPage()
+
+        public IActionResult AddPage()
 		{
 			var categories = _context.Categories.Select(c => new SelectListItem
 			{
@@ -40,14 +51,18 @@ namespace ExpenseTrack.Controllers
 			return View("Add");
 		}
 
-		[HttpPost]
+        [HttpPost]
         public IActionResult AddExpense(Expense expense)
         {
             try
             {
-                    _context.Expenses.Add(expense);
-                    _context.SaveChanges();
-                    return RedirectToAction("Index");
+                var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                expense.UserId = userId;
+
+                _context.Expenses.Add(expense);
+                _context.SaveChanges();
+
+                return RedirectToAction("Index");
             }
             catch (Exception ex)
             {
@@ -56,6 +71,7 @@ namespace ExpenseTrack.Controllers
                 return View("Add", expense);
             }
         }
+
 
 
         public IActionResult EditPage(int ExpenseID)
@@ -82,28 +98,36 @@ namespace ExpenseTrack.Controllers
 
 
 		[HttpPost]
-		public IActionResult EditExpense(Expense expense)
-		{
-			
-				// Update the expense in the database
-				_context.Expenses.Update(expense);
-				_context.SaveChanges();
-				return RedirectToAction("Index"); 
-		}
+        public IActionResult EditExpense(Expense expense)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            expense.UserId = userId;
+
+            // Update the expense in the database
+            _context.Expenses.Update(expense);
+            _context.SaveChanges();
+
+            return RedirectToAction("Index");
+        }
 
 
-		[HttpPost]
-		public IActionResult DeleteExpense(int ExpenseID)
-		{
-			var expense = _context.Expenses.Find(ExpenseID);
 
-			_context.Expenses.Remove(expense);
-			_context.SaveChanges();
+        [HttpPost]
+        public IActionResult DeleteExpense(int ExpenseID)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var expense = _context.Expenses.FirstOrDefault(e => e.ExpenseID == ExpenseID && e.UserId == userId);
+
+            if (expense != null)
+            {
+                _context.Expenses.Remove(expense);
+                _context.SaveChanges();
+            }
+
+            return RedirectToAction("Index");
+        }
 
 
-			return RedirectToAction("Index");
-		}
 
-
-	}
+    }
 }
